+++ packages/passport/sdk-sample-app/src/components/zkevm/EthSignTypedDataV4Examples/etherMailTypedPayload.ts
@@ -1,0 +1,64 @@
+export const getEtherMailTypedPayload = (chainId: number) => ({
+  domain: {
+    name: 'Ether Mail',
+    version: '1',
+    chainId,
+    verifyingContract: '0xd64b0d2d72bb1b3f18046b8a7fc6c9ee6bccd287',
+  },
+  message: {
+    from: {
+      name: 'Cow',
+      wallet: '0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826',
+    },
+    to: {
+      name: 'Bob',
+      wallet: '0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB',
+    },
+    contents: 'Hello, Bob!',
+  },
+  primaryType: 'Mail',
+  types: {
+    EIP712Domain: [
+      {
+        name: 'name',
+        type: 'string',
+      },
+      {
+        name: 'version',
+        type: 'string',
+      },
+      {
+        name: 'chainId',
+        type: 'uint256',
+      },
+      {
+        name: 'verifyingContract',
+        type: 'address',
+      },
+    ],
+    Person: [
+      {
+        name: 'name',
+        type: 'string',
+      },
+      {
+        name: 'wallet',
+        type: 'address',
+      },
+    ],
+    Mail: [
+      {
+        name: 'from',
+        type: 'Person',
+      },
+      {
+        name: 'to',
+        type: 'Person',
+      },
+      {
+        name: 'contents',
+        type: 'string',
+      },
+    ],
+  },
+});