+++ packages/passport/sdk/src/test/mocks.ts
@@ -5,6 +5,10 @@ import { PassportConfiguration } from '../config';
 export const mockErrorMessage = 'Server is down';
 export const mockStarkSignature = 'starkSignature';
 
+export const chainId = 13472;
+export const chainIdHex = '0x34a0';
+export const chainIdEip155 = `eip155:${chainId}`;
+
 const accessToken = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJodHRwczovL2V4YW1wbGUuYXV0aDAuY29tLyIsImF1ZCI6Imh0dHBzOi8vYXBpLmV4YW1wbGUuY29tL2NhbGFuZGFyL3YxLyIsInN1YiI6InVzcl8xMjMiLCJpYXQiOjE0NTg3ODU3OTYsImV4cCI6MTQ1ODg3MjE5Nn0.CA7eaHjIHz5NxeIJoFK9krqaeZrPLwmMmgI_XiQiIkQ';
 
 export const testConfig = new PassportConfiguration({