+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - @imtbl/orderbook: Added `getTrade` and `listTrades` methods for querying trades
+- @imtbl/passport: Added support for `eth_signTypedData_v4` to Passport zkEvm provider. 
 
 ## [0.16.0] - 2023-08-31